Update bed.g - Remove ferences to T1 and T2
There is no such tool as T1 and T2 in the Quad.  It is clear that M3D copied the configs for the quad from one of the other heads and failed to test it.  If they had tested it, at all, then these issues could have easily been fixed by someone that is supposed to be familiar with this code.  Instead, 3+ months after the Quad head was shipping to ProMega backers, I have had to find and fix these errors while I'm trying to just get my printer to work properly for the first time.   I have spent several days worth of hours just trying to get my first print to work properly and have spent so much of that time just fixing things that M3D has failed to delver.

I'm sure support would prefer to not read about my dissatisfaction, but I would prefer to NOT provide my problem solving and technical skills for free.  This unit has had so many defects straight from the manufacturer that it has been completely unusable until I troubleshoot (often with help, sometimes without) a whole host of problems.  I hope that this feedback reached back up into the company since the ways to contact them are so limited.   I would chew out a manager for the state of this project if I could get a hold of one for long enough.   Today is March 3th.  The first Quad Fusion heads were shipped in December (if my understanding is correct).  I received my printer at the end of January.   These problems that I am fixing in your code, uncompensated by you, should have been found and corrected in January or earlier.  Period.