Update _2_Set_Z_Endstop_Height.g
Under ";turn on extruder" section, remove the calls containing P1 and P2.   From what I understand from user DroneOn, the P1 and P2 actually refer to Tool 1 and Tool 2.... not really extruders.   The quad only has one tool:  T0, so P0 is what should be turned on. 

Credit for this fix goes to DroneOn.  I'm just submitting it for everyone else to benefit from it.