Update config.g
I believe this config file was copied over from a different extruder.  For the Quad, there is only 1 extruder heater.  So H2 is nonsense and should be removed.  I also noticed that in the original file, there was only one line for M143 H1 S265  ; Set temperature limit for heater 1 to 265C.   If there really was a second heater, then why wouldn't H2 also have that limit set?

It is my observation that all of these files need to be CAREFULLY reviewed by M3D for errors.  So far I have found many errors. It appears that the quad configuration was copied from a different head and not all the changes were made that were necessary to make it work right, to remove errors, and to remove things that simply don't make sense.