Update config.g
Fixing comment on Drive 0 goes backwards (not forwards).  Also adding D3 to those 3 lines to force Spreadcycle mode all the time so there is no layer shifting caused by the crossover from Stealthchop over to Spreadcycle.  (Thanks DroneOn for explaining this to me and for the fix itself.  I'm merely posting it.  Credit goes to him.)