Update machine_corexy_motions_settings.g
Credit goes to DroneOn.  He was explaining to me that these settings as they were written before were incomplete and that these are closer to what is needed here.  Makes sense to me seeing the obvious dual extruder portion of the line that shouldn't even be there for the Quad...or if it should, it should have 4 ratios, not 1 ratio.   I believe the single value to be correct and is what I'm using now.